+++ src/parser/enrichers/monster/BeastOfTheSea/_module.mjs
@@ -1,0 +1,1 @@
+export { default as BeastsStrike } from "./BeastsStrike.mjs";