+++ src/parser/enrichers/monster/_module.mjs
@@ -25,3 +25,5 @@ export * as Stirge from "./Stirge/_module.mjs";
 export * as SwarmOfStirges from "./SwarmOfStirges/_module.mjs";
 export * as Beholder from "./Beholder/_module.mjs";
 export * as BeastOfTheLand from "./BeastOfTheLand/_module.mjs";
+export * as BeastOfTheSea from "./BeastOfTheSea/_module.mjs";
+export * as BeastOfTheSky from "./BeastOfTheSky/_module.mjs";